+++ src/backend/job-gateway/src/main/java/com/tencent/bk/job/gateway/filter/esb/CheckOpenApiJwtGatewayFilterFactory.java
@@ -25,11 +25,9 @@
 package com.tencent.bk.job.gateway.filter.esb;
 
 import com.tencent.bk.job.common.constant.JobCommonHeaders;
-import com.tencent.bk.job.common.constant.TenantIdConstants;
 import com.tencent.bk.job.common.crypto.util.RSAUtils;
 import com.tencent.bk.job.common.security.autoconfigure.ServiceSecurityProperties;
 import com.tencent.bk.job.common.service.SpringProfile;
-import com.tencent.bk.job.common.tenant.TenantEnvService;
 import com.tencent.bk.job.common.util.JobContextUtil;
 import com.tencent.bk.job.common.util.RequestUtil;
 import com.tencent.bk.job.gateway.model.esb.BkGwJwtInfo;
@@ -55,18 +53,15 @@ public class CheckOpenApiJwtGatewayFilterFactory
     private final SpringProfile springProfile;
     private final ServiceSecurityProperties securityProperties;
 
-    private final TenantEnvService tenantEnvService;
 
     @Autowired
     public CheckOpenApiJwtGatewayFilterFactory(OpenApiJwtService openApiJwtService,
                                                SpringProfile springProfile,
-                                               ServiceSecurityProperties securityProperties,
-                                               TenantEnvService tenantEnvService) {
+                                               ServiceSecurityProperties securityProperties) {
         super(Config.class);
         this.openApiJwtService = openApiJwtService;
         this.springProfile = springProfile;
         this.securityProperties = securityProperties;
-        this.tenantEnvService = tenantEnvService;
     }
 
     @Override
@@ -104,33 +99,11 @@ public class CheckOpenApiJwtGatewayFilterFactory
                 return response.setComplete();
             }
 
-            String tenantId = extractTenantId(request);
-            if (tenantId == null) {
-                response.setStatusCode(HttpStatus.UNAUTHORIZED);
-                return response.setComplete();
-            }
-
             // set header
             request.mutate().header(JobCommonHeaders.APP_CODE, new String[]{authInfo.getAppCode()}).build();
             request.mutate().header(JobCommonHeaders.USERNAME, new String[]{authInfo.getUsername()}).build();
-            request.mutate().header(JobCommonHeaders.BK_TENANT_ID, new String[]{tenantId}).build();
             return chain.filter(exchange.mutate().request(request).build());
         };
-    }
-
-    private String extractTenantId(ServerHttpRequest request) {
-        if (tenantEnvService.isTenantEnabled()) {
-            String tenantId = RequestUtil.getHeaderValue(request, JobCommonHeaders.BK_TENANT_ID);
-            if (StringUtils.isEmpty(tenantId)) {
-                log.error("Missing tenant header from bkApiGateway");
-                return null;
-            } else {
-                return tenantId;
-            }
-        } else {
-            // 如果未开启多租户特性，设置默认租户 default（蓝鲸约定）
-            return TenantIdConstants.DEFAULT_TENANT_ID;
-        }
     }
 
     public boolean validateJwt(BkGwJwtInfo jwtInfo) {
