+++ support-files/kubernetes/charts/bk-job/templates/job-gateway/configmap.yaml
@@ -124,6 +124,7 @@ data:
               predicates:
                 - JobEsbV2Path=/api/job/v2/job-manage/{api_name}
               filters:
+                - CheckTenant
                 - CheckOpenApiJwt
                 - SetPath=/esb/api/v2/{api_name}
                 - RecordEsbAccessLog
@@ -134,6 +135,7 @@ data:
               predicates:
                 - Path=/api/job/v3/job-manage/{api_name}
               filters:
+                - CheckTenant
                 - CheckOpenApiJwt
                 - SetPath=/esb/api/v3/{api_name}
                 - RecordEsbAccessLog
@@ -144,6 +146,7 @@ data:
               predicates:
                 - JobEsbV2Path=/api/job/v2/job-crontab/{api_name}
               filters:
+                - CheckTenant
                 - CheckOpenApiJwt
                 - SetPath=/esb/api/v2/{api_name}
                 - RecordEsbAccessLog
@@ -154,6 +157,7 @@ data:
               predicates:
                 - JobEsbV2Path=/api/job/v3/job-crontab/{api_name}
               filters:
+                - CheckTenant
                 - CheckOpenApiJwt
                 - SetPath=/esb/api/v3/{api_name}
                 - RecordEsbAccessLog
@@ -164,6 +168,7 @@ data:
               predicates:
                 - JobEsbV2Path=/api/job/v2/job-execute/{api_name}
               filters:
+                - CheckTenant
                 - CheckOpenApiJwt
                 - SetPath=/esb/api/v2/{api_name}
                 - RecordEsbAccessLog
@@ -174,6 +179,7 @@ data:
               predicates:
                 - JobEsbV2Path=/api/job/v3/job-execute/{api_name}
               filters:
+                - CheckTenant
                 - CheckOpenApiJwt
                 - SetPath=/esb/api/v3/{api_name}
                 - RecordEsbAccessLog
@@ -184,6 +190,7 @@ data:
               predicates:
                 - JobEsbV2Path=/api/job/v3/job-file-gateway/{api_name}
               filters:
+                - CheckTenant
                 - CheckOpenApiJwt
                 - SetPath=/esb/api/v3/{api_name}
                 - RecordEsbAccessLog
@@ -195,6 +202,7 @@ data:
                 - Path=/iam/api/v1/resources/file_source
               filters:
                 - RecordIamAccessLog
+                - CheckTenant
                 - AddEsbLangHeader
 
             - id: job-ticket-iam
@@ -203,6 +211,7 @@ data:
                 - Path=/iam/api/v1/resources/ticket
               filters:
                 - RecordIamAccessLog
+                - CheckTenant
                 - AddEsbLangHeader
 
             - id: job-analysis-iam
@@ -211,6 +220,7 @@ data:
                 - Path=/iam/api/v1/resources/dashboard_view
               filters:
                 - RecordIamAccessLog
+                - CheckTenant
                 - AddEsbLangHeader
 
             - id: job-manage-iam
@@ -219,6 +229,7 @@ data:
                 - Path=/iam/api/v1/resources/script,/iam/api/v1/resources/task/template,/iam/api/v1/resources/task/plan,/iam/api/v1/resources/account,/iam/api/v1/resources/whitelist,/iam/api/v1/resources/script/public,/iam/api/v1/resources/tag
               filters:
                 - RecordIamAccessLog
+                - CheckTenant
                 - AddEsbLangHeader
 
             - id: job-crontab-iam
@@ -227,6 +238,7 @@ data:
                 - Path=/iam/api/v1/resources/cron/job
               filters:
                 - RecordIamAccessLog
+                - CheckTenant
                 - AddEsbLangHeader
 
             - id: job-execute-iam
@@ -235,6 +247,7 @@ data:
                 - Path=/iam/api/v1/resources/execute/record
               filters:
                 - RecordIamAccessLog
+                - CheckTenant
                 - AddEsbLangHeader
             {{- else if eq .Values.deploy.mode "lite" }}
             - id: job-logout
@@ -267,6 +280,7 @@ data:
               predicates:
                 - JobEsbV2Path=/api/job/v2/{module}/{api_name}
               filters:
+                - CheckTenant
                 - CheckOpenApiJwt
                 - SetPath=/esb/api/v2/{api_name}
                 - RecordEsbAccessLog
@@ -277,6 +291,7 @@ data:
               predicates:
                 - Path=/api/job/v3/{module}/{api_name}
               filters:
+                - CheckTenant
                 - CheckOpenApiJwt
                 - SetPath=/esb/api/v3/{api_name}
                 - RecordEsbAccessLog
@@ -288,6 +303,7 @@ data:
                 - Path=/iam/api/v1/resources/**
               filters:
                 - RecordIamAccessLog
+                - CheckTenant
                 - AddEsbLangHeader
             {{- end }}      
     job:
